+++ examples/with-remix/app/entry.client.tsx
@@ -1,3 +1,4 @@
+import './polyfills';
 import { RemixBrowser } from '@remix-run/react';
 import { hydrate } from 'react-dom';