+++ examples/with-remix/remix.config.js
@@ -7,5 +7,8 @@ module.exports = {
   // assetsBuildDirectory: "public/build",
   // serverBuildPath: "build/index.js",
   // publicPath: "/build/",
-  serverDependenciesToBundle: ['@rainbow-me/rainbowkit'],
+  serverDependenciesToBundle: [
+    '@rainbow-me/rainbowkit',
+    '@rainbow-me/rainbowkit/wallets',
+  ],
 };