+++ packages/rainbowkit/src/wallets/Wallet.ts
@@ -37,6 +37,7 @@ export type Wallet<C extends Connector = Connector> = {
   name: string;
   shortName?: string;
   iconUrl: string | (() => Promise<string>);
+  iconAccent?: string;
   iconBackground: string;
   installed?: boolean;
   downloadUrls?: {
@@ -45,14 +46,22 @@ export type Wallet<C extends Connector = Connector> = {
     browserExtension?: string;
     qrCode?: string;
   };
+  hidden?: (args: {
+    wallets: {
+      id: string;
+      connector: Connector;
+      installed?: boolean;
+    }[];
+  }) => boolean;
   createConnector: () => RainbowKitConnector<C>;
 };
 
 export type WalletList = { groupName: string; wallets: Wallet[] }[];
 
-export type WalletInstance = Omit<Wallet, 'createConnector'> &
+export type WalletInstance = Omit<Wallet, 'createConnector' | 'hidden'> &
   ReturnType<Wallet['createConnector']> & {
     index: number;
+    groupIndex: number;
     groupName: string;
     walletConnectModalConnector?: Connector;
   };