+++ packages/example/CHANGELOG.md
@@ -1,5 +1,46 @@
 # example
 
+## 0.0.36
+
+### Patch Changes
+
+- Updated dependencies [2e6bb8f]
+- Updated dependencies [2e6bb8f]
+- Updated dependencies [2e6bb8f]
+  - @rainbow-me/rainbowkit@0.7.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
+
+## 0.0.35
+
+### Patch Changes
+
+- Updated dependencies [ecd7209]
+- Updated dependencies [248a1cb]
+  - @rainbow-me/rainbowkit@0.6.2
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
+
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [85eb3bd]
+- Updated dependencies [fbf9d82]
+  - @rainbow-me/rainbowkit@0.6.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [c944ddc]
+- Updated dependencies [c944ddc]
+- Updated dependencies [c944ddc]
+- Updated dependencies [62fd332]
+- Updated dependencies [c944ddc]
+- Updated dependencies [c944ddc]
+  - @rainbow-me/rainbowkit@0.6.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.3
+
 ## 0.0.32
 
 ### Patch Changes