+++ packages/rainbowkit-siwe-next-auth/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rainbow-me/rainbowkit-siwe-next-auth
 
+## 0.1.4
+
+### Patch Changes
+
+- 2e6bb8f: Include `v0.7.x` in RainbowKit peer dependency range
+
+## 0.1.3
+
+### Patch Changes
+
+- 62fd332: Include RainbowKit v0.6.x in peer dependency range
+
 ## 0.1.2
 
 ### Patch Changes