+++ packages/rainbowkit/CHANGELOG.md
@@ -1,5 +1,166 @@
 # @rainbow-me/rainbowkit
 
+## 0.7.0
+
+### Minor Changes
+
+- 2e6bb8f: Support tree shaking of individual wallets by exposing them as separate imports via the new `@rainbow-me/rainbowkit/wallets` entry point.
+
+  In order to reduce bundle size, you can now select the individual wallets you want to import into your application.
+
+  Note that since wallets are no longer namespaced via the `wallet` object, all wallets now have a `Wallet` suffix.
+
+  ```diff
+  -import { connectorsForWallets, wallet } from '@rainbow-me/rainbowkit';
+  +import { connectorsForWallets } from '@rainbow-me/rainbowkit';
+  +import {
+  +  injectedWallet,
+  +  rainbowWallet,
+  +  metaMaskWallet,
+  +  coinbaseWallet,
+  +  walletConnectWallet,
+  +} from '@rainbow-me/rainbowkit/wallets';
+
+  const wallets = [
+  -  wallet.injected({ chains }),
+  -  wallet.rainbow({ chains }),
+  -  wallet.metaMask({ chains }),
+  -  wallet.coinbase({ chains, appName: 'My App' }),
+  -  wallet.walletConnect({ chains }),
+  +  injectedWallet({ chains }),
+  +  rainbowWallet({ chains }),
+  +  metaMaskWallet({ chains }),
+  +  coinbaseWallet({ chains, appName: 'My App' }),
+  +  walletConnectWallet({ chains }),
+  ];
+  ```
+
+- 2e6bb8f: The Steakwallet backwards compatibility layer has been removed. Omni should be used instead, available via the new `@rainbow-me/rainbowkit/wallets` entry point.
+
+  ```diff
+  -import { wallet } from '@rainbow-me/rainbowkit';
+  +import { omniWallet } from '@rainbow-me/rainbowkit/wallets';
+
+  const wallets = [
+  -  wallet.steak({ chains }),
+  +  omniWallet({ chains }),
+  ];
+  ```
+
+## 0.6.2
+
+### Patch Changes
+
+- ecd7209: The new Wallet installation flows add two new colors to RainbowKit's `colors` property in the custom `theme` object: `downloadBottomCardBackground` and `downloadTopCardBackground`. Additionally, we now include the optional `iconAccent` prop in the `Wallet` object, specifically for wallets with both a browser extension and an app, to use as the accent color for the new download installation flows.
+- 248a1cb: Generate a new QR code when WalletConnect request is rejected
+
+## 0.6.1
+
+### Patch Changes
+
+- 85eb3bd: Update the desktop Connect Modal to show "Recently Connected Wallets" within their respective group. Previously, they were being plucked into their own "Recent" group.
+- fbf9d82: Ensure `styles.css` directory is included in npm bundle
+
+  This is a follow-up to a fix released in v0.5.2 which added a compatibility layer for Jest's module resolver.
+
+## 0.6.0
+
+### Minor Changes
+
+- c944ddc: Allow custom wallets to automatically hide themselves based on the availability of other wallets in the list. This can be achieved via the new optional `hidden` function on the `Wallet` type.
+
+  **Example usage**
+
+  This is an example of a custom wallet that hides itself if another injected connector is available.
+
+  ```ts
+  import type Wallet from '@rainbow-me/rainbowkit';
+  import { InjectedConnector } from 'wagmi/connectors/injected';
+
+  const myCustomWallet: Wallet = {
+    hidden: ({ wallets }) => {
+      return wallets.some(
+        wallet =>
+          wallet.installed &&
+          (wallet.connector instanceof InjectedConnector ||
+            wallet.id === 'coinbase')
+      );
+    },
+    ...etc,
+  };
+  ```
+
+- c944ddc: Add `installed` property to `wallet.coinbase` to support detecting whether Coinbase Wallet SDK's injected connector is available
+
+### Patch Changes
+
+- c944ddc: Automatically hide "Injected Wallet" option if another injected wallet in the list is available
+
+  **Migration guide**
+
+  Previously we provided instructions for manually calculating whether the "Injected Wallet" option should be visible.
+
+  ```ts
+  const needsInjectedWalletFallback =
+    typeof window !== 'undefined' &&
+    window.ethereum &&
+    !window.ethereum.isMetaMask &&
+    !window.ethereum.isCoinbaseWallet;
+
+  const connectors = connectorsForWallets([
+    {
+      groupName: 'Suggested',
+      wallets: [
+        wallet.rainbow({ chains }),
+        wallet.metaMask({ chains }),
+        wallet.coinbase({ appName: 'My RainbowKit App', chains }),
+        wallet.metaMask({ chains }),
+        ...(needsInjectedWalletFallback ? [wallet.injected({ chains })] : []),
+      ],
+    },
+  ]);
+  ```
+
+  This manual logic should no longer be needed since it's now handled automatically, meaning that the previous example could be simplified in the following way.
+
+  ```ts
+  const connectors = connectorsForWallets([
+    {
+      groupName: 'Suggested',
+      wallets: [
+        wallet.rainbow({ chains }),
+        wallet.metaMask({ chains }),
+        wallet.coinbase({ appName: 'My RainbowKit App', chains }),
+        wallet.metaMask({ chains }),
+        wallet.injected({ chains }),
+      ],
+    },
+  ]);
+  ```
+
+  In addition, since the "Injected Wallet" option is only rendered when necessary based on the end user's browser environment, it's recommended that you place it at the start of the list to ensure it's visible.
+
+  ```ts
+  const connectors = connectorsForWallets([
+    {
+      groupName: 'Suggested',
+      wallets: [
+        wallet.injected({ chains }),
+        wallet.rainbow({ chains }),
+        wallet.metaMask({ chains }),
+        wallet.coinbase({ appName: 'My RainbowKit App', chains }),
+        wallet.metaMask({ chains }),
+      ],
+    },
+  ]);
+  ```
+
+- c944ddc: Move the "Injected Wallet" fallback option to the start of the default wallet list when present
+
+  This option is only presented when an injected wallet is available that isn't handled by another wallet in the list, which means that it's the option most likely to be selected by the end user. As a result, we now give it the highest priority in the list returned from `getDefaultWallets`.
+
+- c944ddc: Ensure TokenPocket is not detected as MetaMask
+
 ## 0.5.3
 
 ### Patch Changes